# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Onfly Travel API",
        "description": "API for managing travel orders and notifications",
        "contact": {
            "email": "support@onfly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "List all travel orders with filters",
                "operationId": "e9f420f189d120781c79e47df24c074f",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by order status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "solicitado",
                                "aprovado",
                                "cancelado"
                            ],
                            "example": "solicitado"
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "description": "Filter by destination (partial match)",
                        "schema": {
                            "type": "string",
                            "example": "São Paulo"
                        }
                    },
                    {
                        "name": "departure_date",
                        "in": "query",
                        "description": "Exact departure date (format: YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-01"
                        }
                    },
                    {
                        "name": "return_date",
                        "in": "query",
                        "description": "Exact return date (format: YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-10"
                        }
                    },
                    {
                        "name": "departure_date_from",
                        "in": "query",
                        "description": "Departure date range start (format: YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-01"
                        }
                    },
                    {
                        "name": "departure_date_to",
                        "in": "query",
                        "description": "Departure date range end (format: YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-31"
                        }
                    },
                    {
                        "name": "return_date_from",
                        "in": "query",
                        "description": "Return date range start (format: YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-10"
                        }
                    },
                    {
                        "name": "return_date_to",
                        "in": "query",
                        "description": "Return date range end (format: YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-31"
                        }
                    },
                    {
                        "name": "created_at_from",
                        "in": "query",
                        "description": "Filter orders created from this date (format: YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-11-01"
                        }
                    },
                    {
                        "name": "created_at_to",
                        "in": "query",
                        "description": "Filter orders created until this date (format: YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-11-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of travel orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TravelOrder"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "Create a new travel order",
                "operationId": "f6da581b524b24641469f160ddc7b019",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTravelOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Travel order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Travel order created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TravelOrder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "Get a specific travel order",
                "operationId": "523101cea33f3f84142a4df32fcc8780",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Travel order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TravelOrder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Travel order not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}/status": {
            "patch": {
                "tags": [
                    "Travel Orders"
                ],
                "summary": "Update travel order status (Admin only)",
                "operationId": "706c2932e14467c3f20ea4adc72bf1e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTravelOrderStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TravelOrder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Travel order not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateTravelOrderRequest": {
                "required": [
                    "requester_name",
                    "destination",
                    "departure_date",
                    "return_date"
                ],
                "properties": {
                    "requester_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "destination": {
                        "type": "string",
                        "example": "New York"
                    },
                    "departure_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-01"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-10"
                    }
                },
                "type": "object"
            },
...
            "UpdateTravelOrderStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "solicitado",
                            "aprovado",
                            "cancelado"
                        ],
                        "example": "solicitado"
                    }
                },
                "type": "object"
            },
            "TravelOrder": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "requester_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "destination": {
                        "type": "string",
                        "example": "New York"
                    },
                    "departure_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-01"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-10"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "solicitado",
                            "aprovado",
                            "cancelado"
                        ],
                        "example": "solicitado"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}